Rename navigation_draver.dart, rename class  NavigationDrawer, remove commented/unused import
