+++ lib/widgets/build_search_text.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-//import 'package:flutter_application_1/pages/main_pages.dart';
 
 Widget buildSearchTextField() => Padding(
       padding: const EdgeInsets.fromLTRB(10, 10, 10, 10),
